ai_service/requirements.txt
# server
fastapi
uvicorn[standard]
httpx

# NLP / parsing
spacy
pdfminer.six
docx2txt

# config & models
pydantic
python-dotenv
